Christopher got the drip
Added two new prototype spritesheets to our boy Christopher: walking to the right and to the left!

I also reversed both of these so now Christopher can point in either direction and walk in either direction, so now he can:

- run to the left and face left
- run to the right and face left
- run to the left and face right
- run to the right and face right

also added new hitbox to Christopher, although this uses a capsule, but that doesn't matter too much but I would like to upgrade that to a poly hitbox in the future.